Trainer

+++ src/resultCache/file_daemon.py
@@ -1,10 +1,10 @@
 from threading import Thread, Event
 from sched import scheduler
-from typing import Union
-import shutil
+from typing import Optional
 import time
 import json
 import os
+import shutil
 
 
 class PurgeDaemon(Thread):
@@ -16,15 +16,26 @@ class PurgeDaemon(Thread):
         self._min_expiry_time = min_expiry_time
         self.file_lock = set()
         self.load_schedule()
+        self.create_data_folder()
+
+    def create_data_folder(self):
+        last_slash = self._schedule_location.rfind("/")
+        if last_slash == -1:
+            return
+
+        path = self._schedule_location[:last_slash]
+        if not os.path.exists(path):
+            os.makedirs(path)
 
     def save_schedule(self):
         to_save = []
         for event in self._scheduler.queue:
             delete_time = (event.argument[1], event.time)
-            to_save.append(delete_time)
+            to_save.append(delete_time)        
 
         with open(self._schedule_location, "w") as file:
             json.dump(to_save, file)
+
 
     def load_schedule(self):
         if not os.path.exists(self._schedule_location):
@@ -143,21 +154,32 @@ class FileHandler:
     def get_expiry_time(self, generation_time) -> int:
         return min(5 * generation_time + self._min_expiry_time, self._max_expiry_time)
 
-    def put(self, username: str, claim_check: str, generation_time: int, data: Union[bytes, str]):
+    def _put_file(self, username: str, claim_check: str, generation_time: int, data: Optional[str], status: float, errors: Optional[str] = None):
         expiry_time = self.get_expiry_time(generation_time) + time.time()
         location = self.get_file_location(username, claim_check, create_folders=True)
+        self._purge_daemon.remove_entry(location)
 
-        mode = "w"
-        if type(data) == bytes:
-            mode = "wb"
+        file_contents = {
+            "data": data,
+            "status": status,   
+            "errors": errors
+        }
 
-        with open(location, mode) as file:
-            file.write(data)
+        with open(location, "w") as file:
+            file.write(json.dumps(file_contents))
 
         self._purge_daemon.add_entry(location, expiry_time)
 
+    def put(self, username: str, claim_check: str, generation_time: int, data: str):
+        self._put_file(username, claim_check, generation_time, data, 1.0)
 
-    def get(self, username: str, claim_check: str, type: type) -> Union[bytes, str]:
+    def update_status(self, username: str, claim_check: str, generation_time: int, status: float):
+        self._put_file(username, claim_check, generation_time, None, status)
+
+    def update_error(self, username: str, claim_check: str, generation_time: int, errors: str):
+        self._put_file(username, claim_check, generation_time, None, 1.0, errors)
+
+    def _get_file(self, username: str, claim_check: str, remove: bool = True) -> dict:
         location = self.get_file_location(username, claim_check)
 
         while location in self._purge_daemon.file_lock:
@@ -168,23 +190,46 @@ class FileHandler:
 
         self._purge_daemon.file_lock.add(location)
 
-        mode = "r"
-        if type == bytes:
-            mode = "rb"
-
-        with open(location, mode) as file:
+        with open(location, "r") as file:
             data = file.read()
 
-        self._purge_daemon.delete(location, True)
-        self._purge_daemon.remove_entry(location)
+        if remove:
+            self._purge_daemon.delete(location, True)
+            self._purge_daemon.remove_entry(location)
+
         self._purge_daemon.file_lock.remove(location)
 
+        return json.loads(data)
 
-        return data
+    def get_status(self, username: str, claim_check: str) -> Optional[float]:
+        file_contents = self._get_file(username, claim_check, False)
+
+        if file_contents == None:
+            return None
+
+        return str(file_contents["status"])
+
+
+    def get(self, username: str, claim_check: str, remove: bool = True) -> Optional[str]:
+        file_contents = self._get_file(username, claim_check, remove=remove)
+
+        if file_contents == None:
+            return None
+
+        result = {}
+
+        if file_contents["errors"] != None:
+            result["return_type"] = "errors"
+            result["contents"] = file_contents["errors"]
+        else:
+            result["return_type"] = "data"
+            result["contents"] = file_contents["data"]
+
+        return json.dumps(result)
 
     def delete_all(self, delete_schedule=False):
         if delete_schedule:
             shutil.rmtree(self._directory, ignore_errors=True)
         else:
             self._purge_daemon.delete_all(False)
-            
+
